+++ vscode/src/edit/provider.ts
@@ -35,9 +35,9 @@ import { DefaultModelParameterProvider } from './adapters/default'
 import { SmartApplyCustomModelParameterProvider } from './adapters/smart-apply-custom'
 import type { EditManagerOptions } from './edit-manager'
 import { responseTransformer } from './output/response-transformer'
-import { DefaultEditPromptBuilder } from './prompt/builder/default'
-import { SmartApplyCustomEditPromptBuilder } from './prompt/builder/smart-apply-custom'
+import { DefaultEditPromptBuilder } from './prompt'
 import { PROMPT_TOPICS, SMART_APPLY_MODEL_IDENTIFIERS } from './prompt/constants'
+import { SmartApplyCustomEditPromptBuilder } from './prompt/smart-apply/apply/smart-apply-custom'
 import type { EditPromptBuilder } from './prompt/type'
 import { EditIntentTelemetryMetadataMapping, EditModeTelemetryMetadataMapping } from './types'
 import { isStreamedIntent } from './utils/edit-intent'
